backend/requirements.txt
openai
python-pptx
requests
python-dotenv
fastapi
uvicorn
python-multipart
